yt-feature/requirements.txt
flask
flask-cors
rich
google-generativeai
pytubefix
bs4
requests
python-dotenv
